instructions/app-requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
requests==2.28.1
python-dotenv==0.19.1
telnyx==1.2.0
assemblyai==0.6.0
openai==1.3.5  # Only needed if using OpenAI for LLM
# anthropic==0.5.0  # Uncomment if using Claude
# llama-cpp-python==0.2.11  # Uncomment if using Llama models locally
redis==4.5.4
uuid==1.30
pathlib==1.0.1
tenacity==8.2.2  # For retries and backoff
pydantic==1.10.8  # For data validation
jsonschema==4.17.3  # For JSON validation
